completed thermalParameter context and Controller
ParametersContext was completed with CRU operations.
thermalParameterControllers was completed with creating a parameter
in case the value is empty.

+++ backend/models/thermalParameterModel.js
@@ -67,7 +67,7 @@ const thermalParameter = (sequelize, DataTypes) => {
         allowNull: false,
       },
       parameter_value: {
-        type: DataTypes.DECIMAL(5, 3),
+        type: DataTypes.FLOAT,
         required: true,
         notNull: true,
         notEmpty: true,
